create CommonButton
