+++ src/App.css
@@ -2,7 +2,7 @@
 * {
   --mainBlue: #1877f2;
   --lightBlue: #5890ff;
-  --lightGrany: #4e4b66;
+  --lightGray: #4e4b66;
   --mainGray: #667080;
   --bgGray: #eef1f4;
   --red: #ec1515;
@@ -149,4 +149,27 @@ a {
   margin-left: 8px;
   vertical-align: top;
   line-height: 24px;
+}
+
+.common-footer {
+  width: 100%;
+  height: 78px;
+  border-top: 2px solid var(--bgGray);
+  margin-top: 15px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.common-footer button {
+  background: var(--bgGray);
+  color: var(--mainGray);
+  border: none;
+  width: 113px;
+  height: 50px;
+  border-radius: 6px;
+  font-weight: bold;
+}
+.common-footer button.active {
+  color: var(--white);
+  background: var(--mainBlue);
 }
