+++ src/components/common/atoms/CommonButton.js
@@ -1,7 +1,12 @@
 import React from "react";
 
-const CommonButton = () => {
-  return <div>CommonButton</div>;
+const CommonButton = ({ commonBtnText }) => {
+  return (
+    <div className="common-footer">
+      <button>{commonBtnText}</button>
+      {/* <button className="active">新規投稿</button> */}
+    </div>
+  );
 };
 
 export default CommonButton;
